+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
   - you can now add custom headers requests (`key: value` format)
   - In diagram, you can visualize `ToOne` relations with the color of the entity on the left.
   - add `requireAuthToken` to admin options to open settings dialog to set the bearer token directly.
+  - add `withLiveQuery` to admin options to control if live query is used. (can be overwritten by local storage settings)
 
 ## [2.7.34] 2025-4-29